chore: move out models from callback folder
